requirements.txt
streamlit==1.31.0
groq==0.4.0
python-dotenv==1.0.0